requirements.txt
Pillow==9.3.0
pynput==1.7.6
pyperclip==1.8.2
pytesseract==0.3.10
